+++ sdk/src/client/bindings/java/examples/src/GenerateAddresses.java
@@ -9,10 +9,12 @@ import org.iota.types.secret.Range;
 public class GenerateAddresses {
     public static void main(String[] args) throws ClientException, InitializeClientException {
         // Build the client.
-        Client client = new Client(new ClientConfig().withNodes(new String[]{"https://api.testnet.shimmer.network"}));
+        Client client = new Client(
+                new ClientConfig().withNodes(new String[] { "https://api.testnet.shimmer.network" }));
 
-        MnemonicSecretManager secretManager = new MnemonicSecretManager("endorse answer radar about source reunion marriage tag sausage weekend frost daring base attack because joke dream slender leisure group reason prepare broken river");
-        
+        MnemonicSecretManager secretManager = new MnemonicSecretManager(
+                "endorse answer radar about source reunion marriage tag sausage weekend frost daring base attack because joke dream slender leisure group reason prepare broken river");
+
         // Generate public address with default account index and range.
         String[] defaultAddresses = client.generateAddresses(secretManager, new GenerateAddressesOptions());
 
@@ -24,7 +26,8 @@ public class GenerateAddresses {
         System.out.println();
 
         // Generate public address with custom account index and range.
-        String[] addresses = client.generateAddresses(secretManager, new GenerateAddressesOptions().withAccountIndex(0).withRange(new Range(0, 4)));
+        String[] addresses = client.generateAddresses(secretManager,
+                new GenerateAddressesOptions().withAccountIndex(0).withRange(new Range(0, 4)));
 
         // Print the addresses.
         System.out.println("List of generated public addresses:");
@@ -34,7 +37,9 @@ public class GenerateAddresses {
         System.out.println();
 
         // Generate internal addresses with custom account index and range.
-        String[] internalAddresses = client.generateAddresses(secretManager, new GenerateAddressesOptions().withInternal(true).withAccountIndex(0).withRange(new Range(0, 4)));
+        String[] internalAddresses = client.generateAddresses(secretManager,
+                new GenerateAddressesOptions().withAccountIndex(0).withRange(new Range(0, 4))
+                        .withOptions(new GenerateAddressesOptions.GenerateAddressOptions().withInternal(true)));
 
         // Print the addresses.
         System.out.println("List of generated internal addresses:");
@@ -43,16 +48,19 @@ public class GenerateAddresses {
         }
         System.out.println();
 
-        // Generate addresses with providing all inputs, that way it can also be done offline without a node.
-        String[] offlineGeneratedAddresses = client.generateAddresses(secretManager, 
-            new GenerateAddressesOptions()
-            .withCoinType(4219)
-            .withAccountIndex(0)
-            .withRange(new Range(0, 4))
-            .withInternal(false)
-            // Generating addresses with client.generateAddresses(secretManager, new GenerateAddressesOptions()), will by default get the bech32_hrp (Bech32
-            // human readable part) from the node info, generating it "offline" requires setting it in the generateAddressesOptions
-            .withBech32Hrp("rms"));
+        // Generate addresses with providing all inputs, that way it can also be done
+        // offline without a node.
+        String[] offlineGeneratedAddresses = client.generateAddresses(secretManager,
+                new GenerateAddressesOptions()
+                        .withCoinType(4219)
+                        .withAccountIndex(0)
+                        .withRange(new Range(0, 4))
+                        .withOptions(new GenerateAddressesOptions.GenerateAddressOptions().withInternal(false))
+                        // Generating addresses with client.generateAddresses(secretManager, new
+                        // GenerateAddressesOptions()), will by default get the bech32_hrp (Bech32
+                        // human readable part) from the node info, generating it "offline" requires
+                        // setting it in the generateAddressesOptions
+                        .withBech32Hrp("rms"));
 
         // Print the addresses.
         System.out.println("List of offline generated public addresses:");