+++ sdk/src/client/builder.rs
@@ -174,8 +174,8 @@ impl ClientBuilder {
         Default::default()
     }
 
-    #[allow(unused_assignments)]
     /// Set the fields from a client JSON config
+    #[allow(unused_assignments)]
     pub fn from_json(mut self, client_config: &str) -> Result<Self> {
         self = serde_json::from_str(client_config)?;
         // validate URLs
@@ -289,8 +289,8 @@ impl ClientBuilder {
     }
 
     /// Sets the amount of workers that should be used for PoW, default is num_cpus::get().
-    pub fn with_pow_worker_count(mut self, worker_count: usize) -> Self {
-        self.pow_worker_count.replace(worker_count);
+    pub fn with_pow_worker_count(mut self, worker_count: impl Into<Option<usize>>) -> Self {
+        self.pow_worker_count = worker_count.into();
         self
     }