+++ sdk/src/client/stronghold/secret.rs
@@ -36,8 +36,7 @@ impl SecretManage for StrongholdAdapter {
         coin_type: u32,
         account_index: u32,
         address_indexes: Range<u32>,
-        internal: bool,
-        _options: Option<GenerateAddressOptions>,
+        options: Option<GenerateAddressOptions>,
     ) -> Result<Vec<Address>> {
         // Prevent the method from being invoked when the key has been cleared from the memory. Do note that Stronghold
         // only asks for a key for reading / writing a snapshot, so without our cached key this method is invocable, but
@@ -53,6 +52,7 @@ impl SecretManage for StrongholdAdapter {
 
         // Addresses to return.
         let mut addresses = Vec::new();
+        let internal = options.map(|o| o.internal).unwrap_or_default();
 
         for address_index in address_indexes {
             let bip_path = vec![HD_WALLET_TYPE, coin_type, account_index, internal as u32, address_index];
@@ -274,7 +274,7 @@ mod tests {
         assert!(Path::new(stronghold_path).exists());
 
         let addresses = stronghold_adapter
-            .generate_addresses(IOTA_COIN_TYPE, 0, 0..1, false, None)
+            .generate_addresses(IOTA_COIN_TYPE, 0, 0..1, None)
             .await
             .unwrap();
 
@@ -310,7 +310,7 @@ mod tests {
         // Address generation returns an error when the key is cleared.
         assert!(
             stronghold_adapter
-                .generate_addresses(IOTA_COIN_TYPE, 0, 0..1, false, None,)
+                .generate_addresses(IOTA_COIN_TYPE, 0, 0..1, None,)
                 .await
                 .is_err()
         );
@@ -319,7 +319,7 @@ mod tests {
 
         // After setting the correct password it works again.
         let addresses = stronghold_adapter
-            .generate_addresses(IOTA_COIN_TYPE, 0, 0..1, false, None)
+            .generate_addresses(IOTA_COIN_TYPE, 0, 0..1, None)
             .await
             .unwrap();