+++ sdk/src/wallet/bindings/java/lib/src/main/java/org/iota/Wallet.java
@@ -20,7 +20,7 @@ import org.iota.types.exceptions.WalletException;
 import org.iota.types.ids.account.AccountAlias;
 import org.iota.types.ids.account.AccountIdentifier;
 import org.iota.types.ids.account.AccountIndex;
-import org.iota.types.account_methods.AddressGenerationOptions;
+import org.iota.types.account_methods.GenerateAddressOptions;
 
 public class Wallet extends NativeApi {
 
@@ -319,14 +319,14 @@ public class Wallet extends NativeApi {
      * @param options The options.
      * @return The generated address.
      */
-    public String generateAddress(int accountIndex, boolean internal, int addressIndex, AddressGenerationOptions.GenerateAddressOptions options, String bechHrp) throws WalletException {
+    public String generateAddress(int accountIndex, int addressIndex,
+            GenerateAddressOptions options, String bechHrp) throws WalletException {
         JsonObject o = new JsonObject();
         o.addProperty("accountIndex", accountIndex);
-        o.addProperty("internal", internal);
         o.addProperty("addressIndex", addressIndex);
         o.add("options", CustomGson.get().toJsonTree(options));
         o.addProperty("bech32Hrp", bechHrp);
-        
+
         return callBaseApi(new WalletCommand("generateAddress", o)).getAsString();
     }
 
@@ -346,7 +346,7 @@ public class Wallet extends NativeApi {
     /**
      * Converts a hex address to a bech32 address.
      *
-     * @param hex The hex address to convert.
+     * @param hex       The hex address to convert.
      * @param bech32Hrp The bech32 human-readable part.
      * @return The bech32 address.
      */